change database to bbolt from pebble
Kept getting weird behaviours with pebble in cache_test.go.
Hence, I decided to switch to bbolt from pebble as it has supports for buckets which eliminates the complexity of adding suffixes to keys as separate "buckets".
Although bbolt is optimised for read heavy operations, bbolt might not be as fast as pebble.
However, since the Cultured Downloader application is read heavy and only writes when downloading new posts, it shouldn't degrade performance too much.